common/functions/SportsAnalytics_Common_Log_Error/requirements.txt
google-cloud-firestore
google-cloud-logging